Add shoryuken client middleware
