+++ lib/request_id/shoryuken.rb
@@ -2,6 +2,7 @@
 
 module Shoryuken
   module Middleware
-    autoload :RequestId, 'shoryuken/middleware/request_id'
+    autoload :RequestId, 'shoryuken/middleware/server/request_id'
+    autoload :RequestId, 'shoryuken/middleware/client/request_id'
   end
 end
